Finish moving CliProperties
