+++ game-core/src/main/java/games/strategy/engine/framework/startup/ui/InGameLobbyWatcher.java
@@ -1,11 +1,11 @@
 package games.strategy.engine.framework.startup.ui;
 
-import static games.strategy.engine.framework.ArgParser.CliProperties.LOBBY_GAME_COMMENTS;
-import static games.strategy.engine.framework.ArgParser.CliProperties.LOBBY_GAME_HOSTED_BY;
-import static games.strategy.engine.framework.ArgParser.CliProperties.LOBBY_HOST;
-import static games.strategy.engine.framework.ArgParser.CliProperties.LOBBY_PORT;
-import static games.strategy.engine.framework.ArgParser.CliProperties.SERVER_PASSWORD;
-import static games.strategy.engine.framework.ArgParser.CliProperties.TRIPLEA_PORT;
+import static games.strategy.engine.framework.CliProperties.LOBBY_GAME_COMMENTS;
+import static games.strategy.engine.framework.CliProperties.LOBBY_GAME_HOSTED_BY;
+import static games.strategy.engine.framework.CliProperties.LOBBY_HOST;
+import static games.strategy.engine.framework.CliProperties.LOBBY_PORT;
+import static games.strategy.engine.framework.CliProperties.SERVER_PASSWORD;
+import static games.strategy.engine.framework.CliProperties.TRIPLEA_PORT;
 
 import java.awt.Frame;
 import java.time.Instant;
